app/requirements.txt
google-genai
fastapi
python-dotenv
uvicorn
markdown2